--- src/components/app/App.css
@@ -1,6 +1,0 @@
-* {
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  margin: 0;
-}
-/*# sourceMappingURL=App.css.map */
