--- src/components/carousel/Carousel.css
@@ -1,20 +1,0 @@
-.carousel {
-  border: 1px solid black;
-  height: 200px;
-  width: 800px;
-  margin: 50px auto;
-}
-
-.carousel ul {
-  list-style: none;
-}
-
-.carousel ul li {
-  float: left;
-}
-
-.carousel ul li img {
-  height: 200px;
-  width: 200px;
-}
-/*# sourceMappingURL=Carousel.css.map */
