+++ src/components/carousel/Carousel.js
@@ -2,47 +2,15 @@ import React, { useEffect, useState } from 'react';
 import CarouselItemSlide from './CarouselItemSlide';
 import './Carousel.scss';
 
-const Carousel = ({ item }) => {
-
-    const [randomItem, setRandomItem] = useState([])
-
-    //    ================ Shuffle function 
-    function shuffle(arr) {
-        for (let i = arr.length - 1; i > 0; i--) {
-            let j = Math.floor(Math.random() * (i + 1));
-            let temp = arr[i]
-            arr[i] = arr[j]
-            arr[j] = temp
-        }
-        return;
-    }
-
-    useEffect(() => {
-        //    ================ Shuffle item 
-        let randomItem = []
-        if (item.length !== 0) {
-            shuffle(item)
-            for (let i = 0; i < 9; i++) {
-                randomItem.push(item[i])
-            }
-        }
-
-        //    ================ update item by useState
-        setRandomItem(randomItem)
-        console.log(randomItem)
-    }, [item])
-
-    useEffect(() => {
-        console.log(randomItem)
-    }, [randomItem])
+const Carousel = (props) => {
 
 
     return (
         <>
             <div className="carousel">
                 <h2>Best Seller</h2>
-                {randomItem.length !== 0 && (
-                        <CarouselItemSlide item={randomItem}/>
+                {props.item.length !== 0 && (
+                        <CarouselItemSlide item={props.item}/>
                 )}
             </div>
             
@@ -51,3 +19,37 @@ const Carousel = ({ item }) => {
 }
 
 export default Carousel
+
+// const [randomItem, setRandomItem] = useState([])
+
+    // //    ================ Shuffle function 
+    // function shuffle(arr) {
+    //     for (let i = arr.length - 1; i > 0; i--) {
+    //         let j = Math.floor(Math.random() * (i + 1));
+    //         let temp = arr[i]
+    //         arr[i] = arr[j]
+    //         arr[j] = temp
+    //     }
+    //     return;
+    // }
+
+    // useEffect(() => {
+    //     //    ================ Shuffle item 
+    //     let randomItemArr = []
+    //     if (item.length !== 0) {
+    //         shuffle(item)
+    //         for (let i = 0; i < 9; i++) {
+    //             randomItemArr.push(item[i])
+    //         }
+    //     }
+
+    //     //    ================ update item by useState
+    //     setRandomItem(randomItemArr)
+    //     // console.log(randomItem)
+    // }, [item])
+
+    // // useEffect(() => {
+    // //     console.log(randomItem)
+    // // }, [randomItem])
+
+    // console.log(randomItem)
