--- src/components/detail/Detail.css
@@ -1,154 +1,0 @@
-* {
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-main {
-  width: 90vw;
-  height: 100vh;
-  margin: 0 auto;
-}
-
-main .detail-container {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  width: 100%;
-}
-
-main .detail-container .price-title,
-main .detail-container .size-title,
-main .detail-container .color-title,
-main .detail-container .quantity-title {
-  font-weight: bold;
-}
-
-main .detail-container .img-wrap {
-  width: 100%;
-  margin-left: 2rem;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  -ms-flex-item-align: start;
-      align-self: flex-start;
-  border: 1px solid #e9dcdc;
-}
-
-main .detail-container .img-wrap img {
-  width: 350px;
-}
-
-main .detail-container .description-wrap {
-  width: 100%;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  margin-right: 2rem;
-  padding: 0 3rem;
-}
-
-main .detail-container .description-wrap .category,
-main .detail-container .description-wrap .product-name,
-main .detail-container .description-wrap .rating-star,
-main .detail-container .description-wrap .price,
-main .detail-container .description-wrap .size-wrap,
-main .detail-container .description-wrap .color-wrap {
-  margin-bottom: 0.5rem;
-}
-
-main .detail-container .description-wrap .category {
-  display: inline-block;
-  -ms-flex-item-align: start;
-      align-self: flex-start;
-  background: #e6e2e2;
-  border-radius: 2px;
-  padding: 3px 4px;
-}
-
-main .detail-container .description-wrap .rating-star {
-  color: #ffe600;
-}
-
-main .detail-container .description-wrap .price {
-  color: #ff0000;
-  font-size: 2.5rem;
-}
-
-main .detail-container .description-wrap .description {
-  text-align: justify;
-  padding-bottom: 1rem;
-  border-bottom: 1.5px solid #cccaca;
-}
-
-main .detail-container .description-wrap .size-title {
-  padding-top: 1rem;
-}
-
-main .detail-container .description-wrap .size-wrap .size-btn {
-  width: 2.5rem;
-  height: 2.5rem;
-  margin: 0.15rem;
-  cursor: pointer;
-  background: #ffffff;
-}
-
-main .detail-container .description-wrap .color-btn {
-  width: 2.5rem;
-  height: 2.5rem;
-  margin: 0.15rem;
-  cursor: pointer;
-  background: #ffffff;
-  color: rgba(255, 255, 255, 0);
-}
-
-main .detail-container .description-wrap .beige {
-  background: #e0cbb5;
-}
-
-main .detail-container .description-wrap .navy {
-  background: #020264;
-}
-
-main .detail-container .description-wrap .gray {
-  background: gray;
-}
-
-main .detail-container .description-wrap .black {
-  background: black;
-}
-
-main .detail-container .description-wrap .brown {
-  background: #723b0e;
-}
-
-main .detail-container .description-wrap .off-white {
-  background: #f8f3f3;
-}
-
-main .detail-container .description-wrap .quantity-wrap {
-  margin-bottom: 2rem;
-}
-
-main .add-to-cart-btn {
-  padding: 0.4rem;
-  font-size: 1.2rem;
-  letter-spacing: 1px;
-  border-radius: 3px;
-  cursor: pointer;
-}
-/*# sourceMappingURL=Detail.css.map */
