--- src/components/footer/Footer.css
@@ -1,8 +1,0 @@
-footer {
-  height: 60px;
-  background-color: #9c9c9c;
-  position: absolute;
-  width: 100%;
-  bottom: 0;
-}
-/*# sourceMappingURL=Footer.css.map */
