--- src/components/header/Header.css
@@ -1,38 +1,0 @@
-nav {
-  position: relative;
-  background-color: #929292;
-  height: 120px;
-}
-
-nav .logo {
-  position: absolute;
-}
-
-nav .searchBar {
-  display: block;
-  margin-left: 50px;
-}
-
-nav .iconWrap {
-  position: absolute;
-  -webkit-transform: translate(0, -50%);
-          transform: translate(0, -50%);
-  top: 50%;
-  right: 0;
-}
-
-nav .iconWrap ul {
-  list-style: none;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-}
-
-nav .iconWrap li {
-  margin: 10px;
-}
-
-nav .iconWrap li:nth-child(1) {
-  margin-right: 100px;
-}
-/*# sourceMappingURL=Header.css.map */
