+++ src/components/searchResult/SearchResult.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const SearchResult = () =>{
+
+}
+
+export default SearchResult
