+++ src/context/products-context.js
@@ -20,7 +20,7 @@ const ProductsProvider = ({ children }) => {
     .then((response) => response.json())
     .then((json) => {
       setProducts(json);
-      console.log(json);
+      // console.log(json);
     })
     .catch((error) => {
       console.error('Error occured', error);
